% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-helpers.R
\name{assert_join_map_levels_by}
\alias{assert_join_map_levels_by}
\title{Check whether the given join_map_levels_by is valid}
\usage{
assert_join_map_levels_by(join_map_levels_by, spdfs_list)
}
\arguments{
\item{join_map_levels_by}{A named vector with the columns to join the map levels by.}

\item{spdfs_list}{A list with the spdfs of all map levels.}
}
\value{
the join_map_levels_by in the right order
}
\description{
The join_map_levels_by must be a named vector of at most one element.
The columns specified in the vector must be data slots of the spdfs in the spdfs_list.
}
