% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-helpers.R
\name{check_gid_columns}
\alias{check_gid_columns}
\title{Check whether the given gid_columns is a valid .}
\usage{
check_gid_columns(gid_columns, spdfs_list)
}
\arguments{
\item{gid_columns}{A named vector to join the map levels by}

\item{spdfs_list}{A list with the spdfs of all map levels}
}
\description{
The gid_columns must be a named vector of at most one element.
The columns specified in the vector must be in the spdfs_list.
}
