% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-helpers.R
\name{check_spdf_list}
\alias{check_spdf_list}
\title{Check whether the given spdf_list is a valid spdf_list and has all the required params.}
\usage{
check_spdf_list(spdfs_list)
}
\arguments{
\item{spdfs_list}{A list with the spdfs of all map levels}
}
\value{
TRUE if spdf_list is valid.
}
\description{
The spdf_list must be a list of at most two elements.
All elements must be a s4 class of type SpatialPolygonsDataFrame.
Every element must have a GID_1
}
