% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp_2014_federal_states}
\alias{gdp_2014_federal_states}
\title{GPD for federal states of Germany for 2014.}
\format{A data frame with 16 rows and 2 variables:
\describe{
  \item{Federal_State}{Name of the federal state}
  \item{GDP_2014}{GDP for the year 2014, in euro}
}}
\source{
Arbeitskreis Volkswirtschaftliche Gesamtrechnungen der Laender:
\url{http://www.deutschlandinzahlen.de}
}
\usage{
gdp_2014_federal_states
}
\description{
A dataset containing the GPD (gross domestic product) for all 16 federal states
of Germany for the year 2014.
}
\keyword{datasets}
