% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leafdown.R
\docType{data}
\name{Leafdown}
\alias{Leafdown}
\title{Leafdown R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Leafdown
}
\description{
Leafdown R6 Class
}
\keyword{datasets}
