% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_election_states}
\alias{us_election_states}
\title{Results of the 2016 US Presidential Election - State Level}
\format{
A data frame with 51 rows and 15 total columns
\describe{
  \item{State}{Name of the State}
  \item{ST}{Abbreviation of the State name}
  \item{Votes}{Total number of votes cast}
  \item{Republicans2016}{Percent of votes for the Republican Party}
  \item{Democrats2016}{Percent of votes for the Democratic Party}
  \item{Green2016}{Percent of votes for the Green Party}
  \item{Libertarians2016}{Percent of votes for the Libertarian Party}
  \item{TotalPopulation}{Total Population of the county}
  \item{Unemployment}{Percent of unemployment}
  \item{White}{Percentage of Whites}
  \item{Black}{Percentage of Blacks}
  \item{Hispanic}{Percentage of Hispanics}
  \item{Asian}{Percentage of Asians}
  \item{Amerindian}{Percentage of Amerindians}
  \item{Other}{Percentage of Other Races}
}
}
\source{
https://github.com/Deleetdk/USA.county.data

Note: The data was aggregated from the county level
}
\usage{
us_election_states
}
\description{
A dataset containing the results of the presidential election and census data (e.g. racial makeup, unemployment)
}
\keyword{datasets}
