% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leafdown.R
\name{Leafdown}
\alias{Leafdown}
\title{Leafdown R6 Class}
\description{
Leafdown R6 Class

Leafdown R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Leafdown$new()}}
\item \href{#method-create_leafdown}{\code{Leafdown$create_leafdown()}}
\item \href{#method-clone}{\code{Leafdown$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Leafdown$new(spdfs_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_leafdown"></a>}}
\subsection{Method \code{create_leafdown()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Leafdown$create_leafdown(map_level = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Leafdown$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
